+++ src/post_treatments/view_generation.cpp
@@ -57,7 +57,7 @@ void print_help(char* prog_path);
 
 
 
-int main( int argc, char* argv[] )
+int main(int argc, char* argv[])
 {
 
 
@@ -67,7 +67,7 @@ int main( int argc, char* argv[] )
   // =================================================================
   //
   // 1) Initialize command-line option variables with default values
-  int32_t gener = 0;
+  int64_t gener = 0;
 
   // 2) Define allowed options
   const char * options_list = "hVg:";
@@ -79,29 +79,29 @@ int main( int argc, char* argv[] )
 
   // 3) Get actual values of the command-line options
   int option;
-  while ( ( option = getopt_long(argc, argv, options_list, long_options_list, NULL) ) != -1 )
+  while ((option = getopt_long(argc, argv, options_list, long_options_list, NULL)) != -1)
   {
-  switch ( option )
+  switch (option)
     {
       case 'h' :
       {
         print_help(argv[0]);
-        exit( EXIT_SUCCESS );
+        exit(EXIT_SUCCESS);
       }
       case 'V' :
       {
         Utils::PrintAevolVersion();
-        exit( EXIT_SUCCESS );
+        exit(EXIT_SUCCESS);
       }
       case 'g' :
       {
-	      if ( strcmp( optarg, "" ) == 0 )
+	      if (strcmp(optarg, "") == 0)
     		{
-    		  printf( "%s: error: Option -g or --gener : missing argument.\n", argv[0] );
-    		  exit( EXIT_FAILURE );
+    		  printf("%s: error: Option -g or --gener : missing argument.\n", argv[0]);
+    		  exit(EXIT_FAILURE);
     		}
 
-	      gener = atol( optarg );
+	      gener = atol(optarg);
         break;
       }
     }
@@ -109,14 +109,14 @@ int main( int argc, char* argv[] )
 
 
 
-  printf( "Displaying generation %" PRId32 "...\n", gener );
+  printf("Displaying generation %" PRId64 "...\n", gener);
 
   // =================================================================
   //                       Read the backup file
   // =================================================================
   // Load simulation from backup
   ae_exp_manager_X11* exp_manager = new ae_exp_manager_X11();
-  exp_manager->load( gener, false, true, false );
+  exp_manager->load(gener, false, true, false);
 
 
 
@@ -127,7 +127,7 @@ int main( int argc, char* argv[] )
 
   exp_manager->toggle_display_on_off();
   exp_manager->display();
-  while ( exp_manager->quit_signal_received() == false )
+  while (exp_manager->quit_signal_received() == false)
   {
     exp_manager->handle_events();
   }
@@ -142,11 +142,11 @@ int main( int argc, char* argv[] )
 
 void print_help(char* prog_name)
 {
-  printf( "\n************* aevol - Artificial Evolution ************* \n\n" );
-  printf( "This program is Free Software. No Warranty.\n" );
-  printf( "Copyright (C) 2009  LIRIS.\n" );
-  printf( "Usage : %s -h\n", prog_name );
-  printf( "   or : %s -f file.ae\n", prog_name );
-  printf( "\t-h : Display this screen\n" );
-  printf( "\t-g or --gener n    : Display the population at generation n\n" );
+  printf("\n************* aevol - Artificial Evolution ************* \n\n");
+  printf("This program is Free Software. No Warranty.\n");
+  printf("Copyright (C) 2009  LIRIS.\n");
+  printf("Usage : %s -h\n", prog_name);
+  printf("   or : %s -f file.ae\n", prog_name);
+  printf("\t-h : Display this screen\n");
+  printf("\t-g or --gener n    : Display the population at generation n\n");
 }
